+++ src/Routes/Login/Login.js
@@ -4,6 +4,7 @@ import Button from '@mui/material/Button';
 import { styled } from '@mui/system';
 import { useNavigate } from 'react-router-dom';
 import { ComponentOne } from '../../Context/ComponentOne';
+import { UserContextProvider } from '../../Context/Context';
 
 export const Login = () =>{
     const navigate = useNavigate()
@@ -16,13 +17,13 @@ export const Login = () =>{
     }
 
     return(
-        <>
+        <UserContextProvider>
         <Column>
          <TextField id="username-id" label="Username" variant="standard" />
          <TextField id="password-id" label="Password" variant="standard" />
          <ButtonStyle variant="contained" onClick={loginHandle}>Login</ButtonStyle>
         </Column>
         <ComponentOne/>
-       </>
+       </UserContextProvider>
     )
 }
